// src/domain/enums/profileEnums.ts
export type ProfileType = "player" | "organization";

export type PlayerLevel =
  | "beginner"
  | "seventh"
  | "sixth"
  | "fifth"
  | "fourth"
  | "third"
  | "second"
  | "first";

export type PlayerPosition = "drive" | "backhand" | "both" | "any";

export type PlayStyle = "recreational" | "competitive" | "both";

export type OrganizationKind = "club" | "tournament" | "group";